feature/MEMO-1-descomposicion-modular
- Se realizó la descomposición modular para así tener la creación de las notas por un lado y la lista de los topicos por otro lado.
- Se tuvo que crear para lograr este cometido un estado global.
